Added Description to all classes Also moved QuadTree.kt and BoundingBox.kt out of the subfolder spatial and removed the folder
