--- src/main/java/de/officeryoda/fhysics/engine/datastructures/ObjectPool.kt
@@ -1,23 +1,0 @@
-package de.officeryoda.fhysics.engine.datastructures
-
-import java.util.*
-
-class ObjectPool<T> {
-
-    private val pool: Stack<T> = Stack()
-
-    @Synchronized
-    fun borrowObject(): T? {
-        return pool.takeIf { it.isNotEmpty() }?.pop()
-    }
-
-    @Synchronized
-    fun returnObject(instance: T) {
-        pool.push(instance)
-    }
-
-    @Synchronized
-    fun clear() {
-        pool.clear()
-    }
-}
