+++ src/main/java/de/officeryoda/fhysics/engine/objects/ConvexPolygon.kt
@@ -3,6 +3,9 @@ package de.officeryoda.fhysics.engine.objects
 import de.officeryoda.fhysics.engine.collision.CollisionInfo
 import de.officeryoda.fhysics.engine.math.Vector2
 
+/**
+ * A ConvexPolygon is a polygon where all interior angles are less than 180 degrees.
+ */
 class ConvexPolygon(
     vertices: Array<Vector2>, // must be CCW and in global space
     angle: Float = 0f,
