+++ src/main/java/de/officeryoda/fhysics/engine/objects/FhysicsObjectType.kt
@@ -1,5 +1,8 @@
 package de.officeryoda.fhysics.engine.objects
 
+/**
+ * The type of [FhysicsObject].
+ */
 enum class FhysicsObjectType {
     CIRCLE,
     RECTANGLE,
